+++ src/app/layout.tsx
@@ -2,9 +2,9 @@ import type { Metadata } from "next";
 import { Montserrat } from "next/font/google";
 import "./globals.css";
 import "react-loading-skeleton/dist/skeleton.css";
-import { DeviseProvider } from "@/components/providers/devise";
+import { DeviseProvider } from "@/components/providers/DeviseProvider";
 import { TanstackProvider } from "@/components/providers/TanstackProvider";
-import { ThemeProvider } from "@/components/providers/theme";
+import { ThemeProvider } from "@/components/providers/ThemeProvider";
 import { HeaderLayout } from "@/components/ui/HeaderLayout";
 import { FooterLayout } from "@/components/ui/FooterLayout";
